Move document line helpers to a dependency-free module and strip a bare final CR
