+++ src/util/yaml-sections-core.ts
@@ -13,13 +13,13 @@
 import { ESPHOME_YAML_INDENT } from "./esphome-yaml-lang.js";
 import { isIndexSegment } from "./nested-values.js";
 import { LIST_SECTIONS } from "./section-entry-overrides.js";
+import { splitYamlDocLines } from "./yaml-doc-lines.js";
 import { readInstanceScalar } from "./yaml-instance-scalars.js";
 import { indentOf, RE_PAIR_LINE, stripComment } from "./yaml-line-walker.js";
 import {
   _skipBlankAndCommentLines,
   endsBlockAtIndent,
   LIST_ITEM_START_RE,
-  splitYamlDocLines,
   TOP_LEVEL_KEY_RE,
 } from "./yaml-section-lexer.js";
